fixed logic error in VFS choose-loose-or-archive-file decision (was causing the archive file to be used regardless of whether the loose one was newer or not)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5994 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/file/vfs/vfs_tree.h
@@ -46,7 +46,7 @@ private:
 	off_t m_size;
 	time_t m_mtime;
 
-	unsigned m_priority;
+	size_t m_priority;
 
 	PIFileLoader m_loader;
 };
